Ensure ssh_known_hosts is present.
If the file ssh_known_hosts is not present it won't be managed
by puppet. This commit adds 'ensure => present' so the file
will be created if it is missing.

It might be a desireable to only 'ensure => present' when
storeconfigs_enabled is set to true. Someone with more puppet
knowledge should do this.