Add support for OpenBSD
The changes to client/config.pp and server/config.pp are because
I use the future parser, just some '' around literal 0s.

Changes to params.pp to support OpenBSD, the sshd_default_params,
UsePAM, is a bit tricky, since OpenBSD sshd is not linked against
pam, therefore, the parameter is not knows to OpenBSDs sshd.
Even overriding it and setting it to 'no' or undef, didn't helped.
If there are other solutions preferred, I'd like to know.

change to hostkeys.pp is because OpenBSD supports ed25519 type
of keys, so that they get added to ssh_known_hosts file.

metadata.json, just added OpenBSD as supported OS.