Purge removed host keys from knownhosts file and add ed25519 support.
An example use case for this is one would disable for example DSA host keys
collected earlier. Disabling DSA is considered a good security practice. To
then actually prevent an attacker to be successful in carrying out a MitM
attack using a DSA key, the public key has to be removed from the client's
ssh_known_hosts file for its trust to be revoked.