Add parameter to manage ssh_known_hosts file
Previously, this module used a file resource to manage the ssh_known_hosts
file to work around an old bug. Some organizations want to manage this file
through their own means and NOT have this module manage the resource. For those
cases, the parameter of $manage_ssh_known_hosts has been added as a boolean
value. If that value does NOT equal the default of true, that file will not
be managed by this module.