Revert "Add explicit allow_virtual parameters"
This reverts commit f63ae3b0b15927625b06e7db093ac661743a9ecf.

`allow_virtual` parameter was added in Puppet 3.6.1. On Puppet version
before this version, this parameter cause a catalogue compilation fail
`Invalid parameter allow_virtual`.

Refs: https://tickets.puppetlabs.com/browse/PUP-2746

This parameter should not be managed inside this module, but on top side
or in puppet configuration file.

  # site.pp or assimiled
  if versioncmp($::puppetversion,'3.6.1') >= 0 {
    $allow_virtual_packages = hiera('allow_virtual_packages',false)
    Package {
      allow_virtual => $allow_virtual_packages,
    }
  }

Or in puppet.conf, in order to remove warnings

  # puppet.conf
  [main]
  disable_warnings=deprecations

Refs: https://docs.puppetlabs.com/references/latest/configuration.html#disablewarnings