Create Hook For Managing SSH Server Service
As a sysadmin, I need to be able to turn the ssh server service off on some systems. These changes allow me to do this in my node manifest files or using my hiera data. 

## How To Use
In hiera, set `$ssh::server::service::enable = 'stopped'` to have puppet stop the ssh service. Set it back to `running` to turn it back on.