added Solaris support
Solaris Secure Shell (aka Sun_SSH) is a fork of OpenSSH.  A few of the default
options are not compatible ... so I followed what was done for OpenBSD.  If
more platform exceptions come along then perhaps sshd_default_options /
ssh_default_options should be part of the earlier case statement ... or
abstracted away elsewhere.

Note: the OpenSSH facts shipped with v2.8.1 are broken and noisy on Solaris.
There are already a number of issues / pull requests related to these
facts not working on non-linux platforms ... so I didn't fix them as part of
this pull request.