Allow disabling knownhosts bug workaround
PUP-1177 was fixed in Puppet 3.7.0, so it's no longer required to manage
this this resource.  This currently conflicts with the puppet-ansible
module which also manages a known_hosts entry.  This allows turning off
management of the resource and also changes it to be an ensure_resource
declaration that is compatible with the one in puppet-ansible.