Adds validate_cmd parameter for sshd config file
Got bit by this today, managed to give a bad config that killed sshd. Which then stopped the service from running, so I was locked out the box! :cry:

`validate_cmd` allows us to check a file with a given test command before it's replaced. So we can check it works before we replace it.

This is false by default, as it requires Puppet 3.6+, plus it could be unexpected behavior.