Allow users to pass a username to defined resource
From the initial commit in 1fbf6a78:
this allows us to use this defined type as often as we want (which is
the purpose of a defined type). Each usage of it creates an entry in
$user/.ssh/config. This is not a breaking change since you can still use
the resource as before. However, you can now call it multiple types by
setting a unique name for it + providing user => $user as a parameter.

Use case: define a block in each puppet profile you have.

Notice: I needed to bump the concat dependency by a major release, this
means that the next release of this software also should be a major one,
even if the API of this module has no breaking changes.

To use this feature it's required to call the defined resource multiple
times. That requires different names for it. We cannot assume that the
name of the defined resource is always a valid user.