Purge and Recurse should be set together (#399)
If you for whatever reason decide to change the value of purge,
recurse should change with it.

This is nothing new and the solution as it's implemented here is the
same I have seen in other popular modules (saz-sudo, ghoneycutt-ssh)
regarding the relation between purge and recurse. If you don't you
can get unexpected behaviour.

Related to #390

Co-authored-by: okopop <okopop@localhost.com>